gan-batik/batikgan-gradio/requirements.txt
torch
numpy
Pillow
gradio
onnxruntime

requests
scipy
psutil
click
imageio
